src/requirements.txt
PyYAML==6.0.1
poetry==1.7.1